bot_service/src/backend/requirements.txt
aiobotocore==2.15.2
aiocache==0.12.3
aiofiles==24.1.0
aiosmtplib==4.0.2
alembic==1.14.0
argon2-cffi==25.1.0
asyncpg==0.30.0
bcrypt==4.0.1
boto3==1.35.36
chardet>=5.2.0
cryptography>=44.0.0
fastapi==0.117.1
fuzzywuzzy==0.18.0
itsdangerous==2.2.0
langfuse==2.27.1
litellm==1.77.5
msoffcrypto-tool>=5.4.2
openai==1.109.1
opensearch-py>=2.7.1
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-sdk==1.29.0
pandas>=2.2.3
passlib==1.7.4
pre-commit==4.0.1
prometheus_client==0.21.1
psycopg2-binary
pydantic>=2.7.4
pyjwt>=2.10.1
pymongo>=4.10.1
PyMuPDF==1.25.1
pyodbc==5.2.0
pypdf==6.0.0
pytest>=8.3.3
pytest-asyncio>=0.24.0
python-dotenv==1.0.0
python-json-logger==3.2.1
python-Levenshtein==0.26.1
python-multipart==0.0.18
python_docx==1.1.2
rich==13.6.0
sqlalchemy==2.0.21
tiktoken>=0.8.0
uvicorn==0.23.2
